Fix first crit every round not being calculated right
