+++ src/crits.cpp
@@ -27,54 +27,6 @@ static int cached_damage    = 0;
 static int crit_damage      = 0;
 static int shot_weapon_mode = 0;
 
-class weapon_info
-{
-public:
-    float crit_bucket{};
-    unsigned int weapon_seed{};
-    unsigned unknown1{};
-    unsigned unknown2{};
-    bool unknown3{};
-    float unknown4{};
-    int crit_attempts{};
-    int crit_count{};
-    float observed_crit_chance{};
-    bool unknown7{};
-    weapon_info()
-    {
-    }
-    void Load(IClientEntity *weapon)
-    {
-        crit_bucket          = *(float *) ((uintptr_t) weapon + 0xa38);
-        weapon_seed          = *(unsigned int *) ((uintptr_t) weapon + 0xb3c);
-        unknown1             = *(unsigned int *) ((uintptr_t) weapon + 0xb30);
-        unknown2             = *(unsigned int *) ((uintptr_t) weapon + 0xb34);
-        unknown3             = *(bool *) ((uintptr_t) weapon + 0xb17);
-        unknown4             = *(float *) ((uintptr_t) weapon + 0xb40);
-        crit_attempts        = *(int *) ((uintptr_t) weapon + 0xa3c);
-        crit_count           = *(int *) ((uintptr_t) weapon + 0xa40);
-        observed_crit_chance = *(float *) ((uintptr_t) weapon + 0xbfc);
-        unknown7             = *(bool *) ((uintptr_t) weapon + 0xb18);
-    }
-    weapon_info(IClientEntity *weapon)
-    {
-        Load(weapon);
-    }
-    void restore_data(IClientEntity *weapon)
-    {
-        *(float *) ((uintptr_t) weapon + 0xa38)        = crit_bucket;
-        *(unsigned int *) ((uintptr_t) weapon + 0xb3c) = weapon_seed;
-        *(unsigned int *) ((uintptr_t) weapon + 0xb30) = unknown1;
-        *(unsigned int *) ((uintptr_t) weapon + 0xb34) = unknown2;
-        *(bool *) ((uintptr_t) weapon + 0xb17)         = unknown3;
-        *(float *) ((uintptr_t) weapon + 0xb40)        = unknown4;
-        *(int *) ((uintptr_t) weapon + 0xa3c)          = crit_attempts;
-        *(int *) ((uintptr_t) weapon + 0xa40)          = crit_count;
-        *(float *) ((uintptr_t) weapon + 0xbfc)        = observed_crit_chance;
-        *(bool *) ((uintptr_t) weapon + 0xb18)         = unknown7;
-    }
-};
-
 static float getWithdrawMult(IClientEntity *wep)
 {
     weapon_info info(wep);
