+++ src/hooks.cpp
@@ -101,6 +101,7 @@ VMTHook baseclientstate{};
 VMTHook baseclientstate8{};
 VMTHook clientmode{};
 VMTHook panel{};
+VMTHook prediction{};
 VMTHook client{};
 VMTHook inventory{};
 VMTHook chathud{};
