+++ src/hooks/CMakeLists.txt
@@ -11,6 +11,7 @@ set(files "${CMAKE_CURRENT_LIST_DIR}/CanPacket.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/others.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/Paint.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/PreDataUpdate.cpp"
+    "${CMAKE_CURRENT_LIST_DIR}/RunCommand.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/SendNetMsg.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/Shutdown.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/FireEvent.cpp"
